Collapse GitHub auth into the onboarding wizard as Step 2
The wizard now wraps the entire first-run flow rather than running
after auth: Welcome → Connect GitHub → Choose repositories → You're
ready. Brand-new users walk all four steps; returning users with a
saved token jump straight to "Choose repositories". Step 2 auto-
advances once an authenticated username resolves.

To avoid duplicating the OAuth/PAT switcher, extract it into a
reusable AuthPanel component on both web and mobile and reuse it from
TokenSetup (now used only as the focused re-auth screen for an
expired token) and the wizard's Connect step.

App routing on both platforms now activates the wizard whenever the
user lacks a token or has no repositories configured. TokenSetup is
reserved for the token-expired case (web) or as a defensive fallback
(both).
